+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.0.2 - 2023-05-11
+
+### Fixed
+
+- Install mapper adaptor will no longer be overriden if already set.
+- Added RequiredRule rules to validator.
+
 ## 1.0.1 - 2023-04-02
 
 ### Fixed